backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
deep-translator==1.11.4
defusedxml==0.7.1
dnspython==2.8.0
edge-tts==7.2.7
email-validator==2.3.0
fastapi==0.135.1
fonttools==4.62.1
fpdf==1.7.2
fpdf2==2.8.7
frozenlist==1.8.0
googletrans==4.0.0rc1
greenlet==3.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.7.1
passlib==1.7.4
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-dotenv==1.2.2
python-multipart==0.0.22
reportlab==4.4.10
requests==2.32.5
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
yarl==1.23.0
